Bugfix: dot naming convention in clip models
